Separate out JSON and wrapper functionality into two a base class and two subclasses
