+++ Sources/JSONSpam.swift
@@ -1,5 +1,5 @@
 //
-//  SwifterSpam.swift
+//  JSONSpam.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
@@ -25,7 +25,7 @@
 
 import Foundation
 
-public extension Swifter {
+public extension SwifterJSON {
 
     /**
     POST   users/report_spam
