+++ Sources/JSONTweets.swift
@@ -1,5 +1,5 @@
 //
-//  SwifterTweets.swift
+//  JSONTweets.swift
 //  Swifter
 //
 //  Copyright (c) 2014 Matt Donnelly.
@@ -25,7 +25,7 @@
 
 import Foundation
 
-public extension Swifter {
+public extension SwifterJSON {
 
     /**
     GET    statuses/retweets/:id
@@ -71,31 +71,6 @@ public extension Swifter {
         parameters["include_entities"] ??= includeEntities
 
         self.getJSON(path: path, baseURL: .api, parameters: parameters, success: { json, _ in success?(json) }, failure: failure)
-    }
-    
-    public func getWrapperTweet(
-        forID id: String,
-        count: Int? = nil,
-        trimUser: Bool? = nil,
-        includeMyRetweet: Bool? = nil,
-        includeEntities: Bool? = nil,
-        success: WrapperSuccessHandler<SwifterTweet>? = nil,
-        failure: FailureHandler? = nil)
-    {
-        let path = "statuses/show.json"
-        
-        var parameters = Dictionary<String, Any>()
-        parameters["id"] = id
-        parameters["count"] ??= count
-        parameters["trim_user"] ??= trimUser
-        parameters["include_my_retweet"] ??= includeMyRetweet
-        parameters["include_entities"] ??= includeEntities
-        
-        self.getWrapper(path: path,
-                        baseURL: .api,
-                        parameters: parameters,
-                        success: { tweet, _ in success?(tweet) },
-                        failure: failure)
     }
 
     /**
