+++ Sources/SwifterJSON.swift
@@ -1,0 +1,27 @@
+//
+//  SwifterJSON.swift
+//  Swifter
+//
+//  Created by Zach Wolfe on 2017-07-23.
+//  Copyright © 2017 Matt Donnelly. All rights reserved.
+//
+
+public class SwifterJSON : SwifterBase {
+    
+    // MARK: - Types
+    
+    public typealias SuccessHandler = (JSON) -> Void
+    public typealias CursorSuccessHandler = (JSON, _ previousCursor: String?, _ nextCursor: String?) -> Void
+    
+    // MARK: - JSON Requests
+    
+    @discardableResult
+    internal func getJSON(path: String, baseURL: TwitterURL, parameters: Dictionary<String, Any>, uploadProgress: HTTPRequest.UploadProgressHandler? = nil, downloadProgress: JSONSuccessHandler? = nil, success: JSONSuccessHandler?, failure: HTTPRequest.FailureHandler?) -> HTTPRequest {
+        return self.jsonRequest(path: path, baseURL: baseURL, method: .GET, parameters: parameters, uploadProgress: uploadProgress, downloadProgress: downloadProgress, success: success, failure: failure)
+    }
+    
+    @discardableResult
+    internal func postJSON(path: String, baseURL: TwitterURL, parameters: Dictionary<String, Any>, uploadProgress: HTTPRequest.UploadProgressHandler? = nil, downloadProgress: JSONSuccessHandler? = nil, success: JSONSuccessHandler?, failure: HTTPRequest.FailureHandler?) -> HTTPRequest {
+        return self.jsonRequest(path: path, baseURL: baseURL, method: .POST, parameters: parameters, uploadProgress: uploadProgress, downloadProgress: downloadProgress, success: success, failure: failure)
+    }
+}
