+++ Sources/Wrapper.swift
@@ -6,4 +6,129 @@
 //  Copyright © 2017 Matt Donnelly. All rights reserved.
 //
 
-protocol SwifterObject: Decodable {}
+import Foundation
+
+public class SwifterWrapper: SwifterBase {
+    
+    // MARK: - Types
+    
+    typealias Object = Decodable
+    
+    public typealias SuccessHandler<T: Object> = (T) -> Void
+    public typealias CursorSuccessHandler<T> =
+        (T, _ previousCursor: String?, _ nextCursor: String?) -> Void
+    where T: Object
+    public typealias WrapperJSONSuccessHandler<T> = (T, _ response: HTTPURLResponse) -> Void
+    where T: Object
+    
+    // MARK: - Wrapper Requests
+    
+    @discardableResult
+    internal func wrapperRequest<T: Object>(
+        path: String,
+        baseURL: TwitterURL,
+        method: HTTPMethodType,
+        parameters: Dictionary<String, Any>,
+        uploadProgress: HTTPRequest.UploadProgressHandler? = nil,
+        // TODO: implement download progress handlers somehow
+        //downloadProgress: ((DownloadProgressData<T>, HTTPURLResponse) -> Void)? = nil,
+        success: WrapperJSONSuccessHandler<T>? = nil,
+        failure: HTTPRequest.FailureHandler? = nil)
+        -> HTTPRequest
+    {
+        /*
+         let jsonDownloadProgressHandler: HTTPRequest.DownloadProgressHandler = { data, _, _, response in
+         
+         guard downloadProgress != nil else { return }
+         
+         guard let jsonResult = try? JSON.parse(jsonData: data) else {
+         let jsonString = String(data: data, encoding: .utf8)
+         let jsonChunks = jsonString!.components(separatedBy: "\r\n")
+         
+         for chunk in jsonChunks where !chunk.utf16.isEmpty {
+         guard let chunkData = chunk.data(using: .utf8), let jsonResult = try? JSON.parse(jsonData: chunkData) else { continue }
+         downloadProgress?(jsonResult, response)
+         }
+         return
+         }
+         
+         downloadProgress?(jsonResult, response)
+         }*/
+        
+        let jsonSuccessHandler: HTTPRequest.SuccessHandler = { data, response in
+            DispatchQueue.global(qos: .utility).async {
+                do {
+                    let decoder = JSONDecoder()
+                    let result = try decoder.decode(T.self, from: data)
+                    DispatchQueue.main.async {
+                        success?(result, response)
+                    }
+                } catch {
+                    DispatchQueue.main.async {
+                        failure?(error)
+                    }
+                }
+            }
+        }
+        
+        if method == .POST {
+            return self.client.get(path,
+                                   baseURL: baseURL,
+                                   parameters: parameters,
+                                   uploadProgress: uploadProgress,
+                                   downloadProgress: nil,
+                                   success: jsonSuccessHandler,
+                                   failure: failure)
+        } else {
+            return self.client.post(path,
+                                    baseURL: baseURL,
+                                    parameters: parameters,
+                                    uploadProgress: uploadProgress,
+                                    downloadProgress: nil,
+                                    success: jsonSuccessHandler,
+                                    failure: failure)
+        }
+    }
+    
+    @discardableResult
+    internal func getWrapper<T: Object>(
+        path: String,
+        baseURL: TwitterURL,
+        parameters: Dictionary<String, Any>,
+        uploadProgress: HTTPRequest.UploadProgressHandler? = nil,
+        // see above
+        //downloadProgress: JSONSuccessHandler? = nil,
+        success: WrapperJSONSuccessHandler<T>?,
+        failure: HTTPRequest.FailureHandler?)
+        -> HTTPRequest
+    {
+        return self.wrapperRequest(path: path,
+                                   baseURL: baseURL,
+                                   method: .GET,
+                                   parameters: parameters,
+                                   uploadProgress: uploadProgress,
+                                   success: success,
+                                   failure: failure)
+    }
+    
+    @discardableResult
+    internal func postWrapper<T: Object>(
+        path: String,
+        baseURL: TwitterURL,
+        parameters: Dictionary<String, Any>,
+        uploadProgress: HTTPRequest.UploadProgressHandler? = nil,
+        // see above
+        //downloadProgress: JSONSuccessHandler? = nil,
+        success: WrapperJSONSuccessHandler<T>?,
+        failure: HTTPRequest.FailureHandler?)
+        -> HTTPRequest
+    {
+        return self.wrapperRequest(path: path,
+                                   baseURL: baseURL,
+                                   method: .POST,
+                                   parameters: parameters,
+                                   uploadProgress: uploadProgress,
+                                   success: success,
+                                   failure: failure)
+    }
+}
