--- Sources/WrapperUser.swift
@@ -1,23 +1,0 @@
-//
-//  WrapperUser.swift
-//  Swifter
-//
-//  Created by Zach Wolfe on 2017-05-18.
-//  Copyright © 2017 Matt Donnelly. All rights reserved.
-//
-
-import Foundation
-
-public struct SwifterUser: SwifterObject {
-    public let profileImageURL: URL
-    public let profileImageURLHTTPS: URL
-    public let name: String
-    public let screenName: String
-    
-    enum CodingKeys: String, CodingKey {
-        case profileImageURL = "profile_image_url"
-        case profileImageURLHTTPS = "profile_image_url_https"
-        case name = "name"
-        case screenName = "screen_name"
-    }
-}
